Data.Default.Instances.Base

Documentation

This module defines Default instances for the types Int, Int8, Int16, Int32, Int64, Word, Word8, Word16, Word32, Word64, Integer, Float, Double, Ratio, Complex, '(->)', IO, Maybe, '()', '[]', Ordering, Any, All, Last, First, Sum, Product, Endo, Dual, and tuples.